# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AiChatMessageRole =
  (typeof AiChatMessageRole)[keyof typeof AiChatMessageRole];

export const AiChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface AiChatMessage {
  role: AiChatMessageRole;
  content: string;
}

export interface AiChatRequest {
  /** Optional system prompt prepended to the conversation. */
  system?: string;
  /** @minItems 1 */
  messages: AiChatMessage[];
  /**
   * @minimum 1
   * @maximum 8192
   */
  maxTokens?: number;
  /** Optional model override. Defaults to the server-configured model. */
  model?: string;
  /** When true, gives the model access to Anthropic's web_search tool (single use). */
  webSearch?: boolean;
}

export interface AiChatResult {
  text: string;
  model: string;
  stopReason?: string | null;
}

export interface ErrorResponse {
  error: string;
}